Improve label perf by caching inner AttributedLabel
