+++ BlueprintUICommonControls/Sources/Label.swift
@@ -6,31 +6,78 @@ import UIKit
 public struct Label: ProxyElement {
 
     /// The text to be displayed.
-    public var text: String
+    public var text: String {
+        didSet { self.updateBackingLabel() }
+    }
     
-    public var font: UIFont = UIFont.systemFont(ofSize: UIFont.systemFontSize)
-    public var color: UIColor = .black
-    public var alignment: NSTextAlignment = .left
-    public var numberOfLines: Int = 0
-    public var lineBreakMode: NSLineBreakMode = .byWordWrapping
-    public var lineHeight: LineHeight = .font
-    public var isAccessibilityElement = false
+    public var font: UIFont = UIFont.systemFont(ofSize: UIFont.systemFontSize) {
+        didSet { self.updateBackingLabel() }
+    }
+    
+    public var color: UIColor = .black {
+        didSet { self.updateBackingLabel() }
+    }
+    
+    public var alignment: NSTextAlignment = .left {
+        didSet { self.updateBackingLabel() }
+    }
+    
+    public var numberOfLines: Int = 0 {
+        didSet { self.updateBackingLabel() }
+    }
+    
+    public var lineBreakMode: NSLineBreakMode = .byWordWrapping {
+        didSet { self.updateBackingLabel() }
+    }
+    
+    public var lineHeight: LineHeight = .font {
+        didSet { self.updateBackingLabel() }
+    }
+    
+    public var isAccessibilityElement = false {
+        didSet { self.updateBackingLabel() }
+    }
+    
+    private var backingLabel : AttributedLabel? = nil
+    private var finishedInit : Bool = false
 
     public init(text: String, configure: (inout Label) -> Void = { _ in }) {
         self.text = text
         configure(&self)
+        self.finishedInit = true
+        self.updateBackingLabel()
     }
 
-    private var effectiveLineBreakMode: NSLineBreakMode {
-        // These line break modes don't work when numberOfLines is 1, and they break line height adjustments.
-        // Normalize them to clipping mode instead.
-        if numberOfLines == 1 && (lineBreakMode == .byCharWrapping || lineBreakMode == .byWordWrapping) {
-            return .byClipping
-        }
-        return lineBreakMode
+    public var elementRepresentation: Element {
+        self.backingLabel ?? self.makeBackingLabel()
     }
     
-    private var attributedText: NSAttributedString {
+    private mutating func updateBackingLabel() {
+        self.backingLabel = self.makeBackingLabel()
+    }
+    
+    private func makeBackingLabel() -> AttributedLabel {
+        AttributedLabel(attributedText: self.makeAttributedText()) { label in
+            label.numberOfLines = numberOfLines
+            label.isAccessibilityElement = isAccessibilityElement
+
+            switch lineHeight {
+            case .custom(let lineHeight, .top):
+                let leading = lineHeight - font.lineHeight
+                label.textRectOffset = UIOffset(horizontal: 0, vertical: -leading)
+
+            case .custom(let lineHeight, .center):
+                let halfLeading = (lineHeight - font.lineHeight) / 2
+                label.textRectOffset = UIOffset(horizontal: 0, vertical: -halfLeading)
+
+            case .font, .custom(_, .bottom):
+                // do nothing, use default behavior
+                break
+            }
+        }
+    }
+    
+    private func makeAttributedText() -> NSAttributedString {
         let paragraphStyle = NSMutableParagraphStyle()
         paragraphStyle.alignment = alignment
         paragraphStyle.lineBreakMode = effectiveLineBreakMode
@@ -51,28 +98,17 @@ public struct Label: ProxyElement {
                 NSAttributedString.Key.font: font,
                 NSAttributedString.Key.foregroundColor: color,
                 NSAttributedString.Key.paragraphStyle: paragraphStyle
-            ])
+            ]
+        )
     }
-
-    public var elementRepresentation: Element {
-        AttributedLabel(attributedText: attributedText) { label in
-            label.numberOfLines = numberOfLines
-            label.isAccessibilityElement = isAccessibilityElement
-
-            switch lineHeight {
-            case .custom(let lineHeight, .top):
-                let leading = lineHeight - font.lineHeight
-                label.textRectOffset = UIOffset(horizontal: 0, vertical: -leading)
-
-            case .custom(let lineHeight, .center):
-                let halfLeading = (lineHeight - font.lineHeight) / 2
-                label.textRectOffset = UIOffset(horizontal: 0, vertical: -halfLeading)
-
-            case .font, .custom(_, .bottom):
-                // do nothing, use default behavior
-                break
-            }
+    
+    private var effectiveLineBreakMode: NSLineBreakMode {
+        // These line break modes don't work when numberOfLines is 1, and they break line height adjustments.
+        // Normalize them to clipping mode instead.
+        if numberOfLines == 1 && (lineBreakMode == .byCharWrapping || lineBreakMode == .byWordWrapping) {
+            return .byClipping
         }
+        return lineBreakMode
     }
 }
